spi: omap2-mcspi: In case of dma errors fall back to pio
In case there are dma errors currently the driver exits.
Make the spi driver fall back to pio mode in case of dma errors.

If the DMA engine is not selected the driver
exits.This patch makes the spi fall back to pio in that case.

Also adds a field dma_unusable to struct omap2_mcspi.

Signed-off-by: Shubhrajyoti D <shubhrajyoti@ti.com>

Conflicts:

	drivers/spi/spi-omap2-mcspi.c

Conflicts:

	drivers/spi/spi-omap2-mcspi.c